requirements.txt
Flask==1.1.2
numpy==1.18.5
tensorflow==2.3.0
Pillow==7.2.0
mysql-connector-python==8.0.21
pandas==1.1.1
SQLAlchemy==1.3.19
opencv-python==4.3.0.36
mtcnn==0.1.0
sklearn==0.0
PyMySQL==0.10.0